feat: simplify tasks
Creating a table proved to be very challenging for people that have
never touched HTML before. As such, we change the tasks to require a
list instead, so as to avoid unnecessary complexity and save time.

Additionally, the same can be said for having multiple inputs. To reduce
complexity and save time, we also remove all other fields but colleague
name.